src/touri_perception/requirements.txt
absl-py==1.2.0
actionlib==1.13.2
addict==2.4.0
angles==1.9.13
antlr4-python3-runtime==4.9.3
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
attrs==22.1.0
Babel==2.10.3
backcall==0.2.0
base-local-planner==1.17.2
beautifulsoup4==4.11.1
black==22.3.0
bleach==5.0.1
bondpy==1.8.6
cachetools==5.2.0
camera-calibration==1.16.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
catkin-pkg==0.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
cma==3.2.2
colorama==0.4.5
contourpy==1.0.5
controller-manager==0.19.5
controller-manager-msgs==0.19.5
cv-bridge==1.16.1
cycler==0.11.0
ddynamic-reconfigure-python==0.0.1
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
distro==1.8.0
docutils==0.19
drawnow==0.72.5
dynamic-reconfigure==1.7.3
dynamixel-sdk==3.7.31
empy==3.3.4
entrypoints==0.4
executing==1.1.1
fairscale==0.4.11
fastjsonschema==2.16.2
fonttools==4.37.4
freetype-py==2.3.0
future==0.18.2
fvcore==0.1.5.post20220512
gazebo-plugins==2.9.2
gazebo-ros==2.9.2
gencpp==0.6.5
geneus==3.0.0
genlisp==0.4.18
genmsg==0.5.16
gennodejs==2.0.2
genpy==0.6.15
gitdb==4.0.9
GitPython==3.1.29
google-auth==2.12.0
google-auth-oauthlib==0.4.6
grpcio==1.49.1
hello-robot-stretch-body==0.4.9
hello-robot-stretch-body-tools==0.4.5
hello-robot-stretch-factory==0.3.9
hello-robot-stretch-tool-share==0.2.5
hydra-core==1.2.0
idna==3.4
image-geometry==1.16.1
imageio==2.22.1
importlib-metadata==5.0.0
importlib-resources==5.9.0
inputs==0.5
interactive-markers==1.12.0
iopath==0.1.9
ipykernel==6.16.0
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
json5==0.9.10
jsonschema==4.16.0
jupyter==1.0.0
jupyter-client==7.3.5
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter-packaging==0.12.3
jupyter-server==1.19.1
jupyterlab==3.4.8
jupyterlab-pygments==0.2.2
jupyterlab-server==2.15.2
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
laser-geometry==1.6.7
llvmlite==0.39.1
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
message-filters==1.15.14
mistune==2.0.4
moveit-core==1.1.9
moveit-ros-planning-interface==1.1.9
moveit-ros-visualization==1.1.9
mpmath==1.2.1
msgpack==1.0.4
mypy-extensions==0.4.3
nbclassic==0.4.5
nbclient==0.7.0
nbconvert==7.2.1
nbformat==5.7.0
nest-asyncio==1.5.6
netifaces==0.11.0
networkx==2.2
nose==1.3.7
notebook==6.4.12
notebook-shim==0.1.0
numba==0.56.2
numpy==1.23.3
oauthlib==3.2.1
omegaconf==2.2.3
open3d==0.13.0
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
packaging==21.3
pandas==1.5.0
pandocfilters==1.5.0
parso==0.8.3
pathlib==1.0.1
pathspec==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pixel-ring==0.1.0
pkgutil-resolve-name==1.3.10
platformdirs==2.5.2
portalocker==2.5.1
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.5
pycollada==0.6
pycparser==2.21
pydot==1.4.2
pyglet==1.5.27
Pygments==2.13.0
PyOpenGL==3.1.0
pyparsing==3.0.9
pyrealsense2==2.51.1.4348
pyrender==0.1.45
pyrsistent==0.18.1
pyserial==3.5
python-dateutil==2.8.2
python-qt-binding==0.4.4
pytz==2022.4
pyusb==1.2.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
qtconsole==5.3.2
QtPy==2.2.1
renamed-opencv-python-inference-engine==2022.1.5
requests==2.28.1
requests-oauthlib==1.3.1
resource-retriever==1.12.7
ros-numpy==0.0.5
rosbag==1.15.14
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosdistro==0.9.0
rosgraph==1.15.14
rosinstall==0.7.8
roslaunch==1.15.14
roslib==1.15.8
roslint==0.12.0
roslz4==1.15.14
rosmake==1.15.8
rosmaster==1.15.14
rosmsg==1.15.14
rosnode==1.15.14
rosparam==1.15.14
rospkg==1.4.0
rospy==1.15.14
rosservice==1.15.14
rostest==1.15.14
rostopic==1.15.14
rosunit==1.15.8
roswtf==1.15.14
rplidar-roboticia==0.9.5
rqt-action==0.4.9
rqt-bag==0.5.1
rqt-bag-plugins==0.5.1
rqt-console==0.4.11
rqt-dep==0.4.12
rqt-graph==0.4.14
rqt-gui==0.5.3
rqt-gui-py==0.5.3
rqt-image-view==0.4.16
rqt-launch==0.4.9
rqt-logger-level==0.4.11
rqt-moveit==0.5.10
rqt-msg==0.4.10
rqt-nav-view==0.5.7
rqt-plot==0.4.13
rqt-pose-view==0.5.11
rqt-publisher==0.4.10
rqt-py-common==0.5.3
rqt-py-console==0.4.10
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.14
rqt-robot-steering==0.5.12
rqt-runtime-monitor==0.5.9
rqt-rviz==0.7.0
rqt-service-caller==0.4.10
rqt-shell==0.4.11
rqt-srv==0.4.9
rqt-tf-tree==0.6.3
rqt-top==0.4.10
rqt-topic==0.4.13
rqt-web==0.4.10
rsa==4.9
rtabmap-ros==0.20.18
rviz==1.14.19
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.2
Send2Trash==1.8.0
sensor-msgs==1.13.1
six==1.16.0
smach==2.5.0
smach-ros==2.5.0
smclib==1.8.6
smmap==5.0.0
snakeviz==2.1.1
sniffio==1.3.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
spidev==3.5
srdfdom==0.6.3
stack-data==0.5.1
sympy==1.11.1
tabulate==0.8.10
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.0.1
terminado==0.16.0
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.5
tf2-kdl==0.7.5
tf2-py==0.7.5
tf2-ros==0.7.5
threadpoolctl==3.1.0
tifffile==2022.8.12
timm==0.6.7
tinycss2==1.1.1
tomli==2.0.1
tomlkit==0.11.5
topic-tools==1.15.14
torch==1.12.1
torchvision==0.13.1
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
trimesh==3.6.38
typing-extensions==4.3.0
urdfdom-py==0.4.6
urdfpy==0.0.22
urllib3==1.26.12
vcstools==0.1.42
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wstool==0.1.17
xacro==1.14.13
yacs==0.1.8
zipp==3.8.1
